+++ core/utils.py
@@ -1,7 +1,6 @@
-import uuid
 import json
-
-from typing import SupportsRound, Any
+import uuid
+from typing import Any, SupportsRound
 
 
 def round_half_up(number: SupportsRound[Any], ndigits: int = 0):
@@ -21,6 +20,10 @@ def is_valid_uuid(val):
         return True
     except ValueError:
         return False
+
+
+def get_uuid_or_none(uuid_or_string):
+    return is_valid_uuid(uuid_or_string)
 
 
 def get_key_from_json_field(json_field, key, default=""):
